requirements.txt
flask
qrcode[pil]
Pillow
requests
brother_ql
